% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-other.R
\docType{data}
\name{BlogPosts}
\alias{BlogPosts}
\title{Tiny 'quanteda' corpus of Japanese blog posts}
\format{
An object of class \code{corpus} (inherits from \code{character}) of length 20.
}
\source{
\url{https://github.com/paithiov909/documents/tree/master/articles}
}
\usage{
BlogPosts
}
\description{
Tiny 'quanteda' corpus of Japanese blog posts
}
\keyword{datasets}
