% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_emoji}
\alias{clean_emoji}
\title{Remove emojis}
\usage{
clean_emoji(text, replacement = "")
}
\arguments{
\item{text}{A character vector.}

\item{replacement}{String.}
}
\value{
A character vector.
}
\description{
Remove emojis
}
