% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_url}
\alias{clean_url}
\title{Remove URLs}
\usage{
clean_url(text, replacement = "")
}
\arguments{
\item{text}{A character vector.}

\item{replacement}{String.}
}
\value{
A character vector.
}
\description{
Remove URLs
}
