% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldccr-data.R
\docType{data}
\name{GimeiAddresses}
\alias{GimeiAddresses}
\title{Japanese Dummy Addresses}
\format{
An object of class \code{data.frame} with 3076 rows and 4 columns.
}
\source{
\url{https://github.com/willnet/gimei/blob/main/lib/data/addresses.yml}
}
\usage{
GimeiAddresses
}
\description{
Japanese Dummy Addresses
}
\keyword{datasets}
