% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-gimei.R
\docType{data}
\name{GimeiNames}
\alias{GimeiNames}
\title{Japanese Dummy Names}
\format{
An object of class \code{data.frame} with 13739 rows and 5 columns.
}
\source{
\url{https://github.com/willnet/gimei/blob/main/lib/data/names.yml}
}
\usage{
GimeiNames
}
\description{
Japanese Dummy Names
}
\keyword{datasets}
