% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_censored}
\alias{is_censored}
\title{Ensure if characters do not include certain characters}
\usage{
is_censored(text, patterns)
}
\arguments{
\item{text}{A character vector.}

\item{patterns}{A character vector; patterns that should be censored.}
}
\value{
Logicals.
}
\description{
Ensure if characters do not include certain characters
}
