% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_within_era}
\alias{is_within_era}
\title{Check if dates are within Japanese era}
\usage{
is_within_era(date, era)
}
\arguments{
\item{date}{Dates.}

\item{era}{String.}
}
\value{
Logicals.
}
\description{
Check if dates are within Japanese era
}
