% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrte-reader.R
\name{jrte_rte_files}
\alias{jrte_rte_files}
\title{Data for Textual Entailment}
\usage{
jrte_rte_files(
  keep = c("rte.nlp2020_base", "rte.nlp2020_append", "rte.lrec2020_surf",
    "rte.lrec2020_sem_short", "rte.lrec2020_sem_long", "rte.lrec2020_me")
)
}
\arguments{
\item{keep}{Character vector. File names to parse.}
}
\value{
tsv file names.
}
\description{
Data for Textual Entailment
}
\seealso{
Other jrte-reader:
\code{\link[=parse_jrte_judges]{parse_jrte_judges()}},
\code{\link[=parse_jrte_reasoning]{parse_jrte_reasoning()}},
\code{\link[=read_jrte]{read_jrte()}}
}
\concept{jrte-reader}
