% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldcc_parser.R
\docType{data}
\name{ldcc_categories}
\alias{ldcc_categories}
\title{List of Categories}
\format{
An object of class \code{list} of length 9.
}
\usage{
ldcc_categories
}
\description{
List of Categories
}
\keyword{datasets}
