% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ldcc.R
\name{ldcc_jnormalize}
\alias{ldcc_jnormalize}
\title{Convert characters following the rules of 'neologd'}
\usage{
ldcc_jnormalize(str)
}
\arguments{
\item{str}{Input vector.}
}
\value{
a character
}
\description{
Convert the characters into normalized style
basing on rules that is recommended for using the Neologism dictionary for MeCab.
}
\seealso{
\url{https://github.com/neologd/mecab-ipadic-neologd/wiki/Regexp.ja}
}
