% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldccr-package.R
\docType{package}
\name{ldccr}
\alias{ldccr}
\alias{ldccr-package}
\title{Japanese Corpus Parser for R}
\description{
Download Livedoor News Corpus, parse to data.frame and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/ldccr}
  \item Report bugs at \url{https://github.com/paithiov909/ldccr/issues}
}

}
\author{
\strong{Maintainer}: Kato Akiru \email{paithiov909@gmail.com}

}
\keyword{internal}
