% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MKWorks}
\alias{MKWorks}
\title{Some Shinji-Shinkana works written by Miyazawa Kenji from Aozora Bunko}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 874 rows and 2 columns.
}
\usage{
MKWorks
}
\description{
Some Shinji-Shinkana works written by Miyazawa Kenji from Aozora Bunko
}
\keyword{datasets}
