% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldccr-data.R
\docType{data}
\name{NekoText}
\alias{NekoText}
\title{Whole text of ‘Wagahai Wa Neko Dearu’ written by Natsume Souseki from Aozora Bunko}
\format{
An object of class \code{character} of length 2258.
}
\source{
\url{https://www.aozora.gr.jp/cards/000148/files/789_ruby_5639.zip}
}
\usage{
NekoText
}
\description{
Whole text of ‘Wagahai Wa Neko Dearu’ written by Natsume Souseki from Aozora Bunko
}
\keyword{datasets}
