% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrte-parse.R
\name{parse_jrte_judges}
\alias{parse_jrte_judges}
\title{Parse judges column of 'rte.*.tsv'}
\usage{
parse_jrte_judges(tbl)
}
\arguments{
\item{tbl}{A tibble returned from \code{\link[=read_jrte]{read_jrte()}} of which name is \code{rte.*.tsv}
(this function cannot parse judges of the \code{pn} dataset).}
}
\value{
A tibble.
}
\description{
Parse judges column of 'rte.*.tsv'
}
\seealso{
Other jrte-reader:
\code{\link[=jrte_rte_files]{jrte_rte_files()}},
\code{\link[=parse_jrte_reasoning]{parse_jrte_reasoning()}},
\code{\link[=read_jrte]{read_jrte()}}
}
\concept{jrte-reader}
\keyword{internal}
