% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldcc_parser.R
\name{parse_ldcc}
\alias{parse_ldcc}
\title{Livedoor News Corpus Parser}
\usage{
parse_ldcc(
  url = "https://www.rondhuit.com/download/ldcc-20140209.tar.gz",
  exdir = tempdir(),
  keep = ldcc_categories,
  collapse = "\\n\\n"
)
}
\arguments{
\item{url}{string.}

\item{exdir}{string. path to save text files.}

\item{keep}{list. categories to parse and keep in data.frame.}

\item{collapse}{string with which \code{paste()} collapses lines.}
}
\value{
data.frame
}
\description{
Download Livedoor News Corpus and parse to data.frame.
}
\seealso{
\url{https://www.rondhuit.com/download.html#ldcc}
}
