% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ja-text8-reader.R
\name{read_ja_text8}
\alias{read_ja_text8}
\title{Read the ja.text8 corpus}
\usage{
read_ja_text8(
 
    url = "https://s3-ap-northeast-1.amazonaws.com/dev.tech-sketch.jp/chakki/public/ja.text8.zip",
  size = NULL
)
}
\arguments{
\item{url}{String.}

\item{size}{Integer. If supplied, samples rows by this argument.}
}
\description{
Download and read the ja.text8 corpus as a tibble.
}
\details{
By default, this function reads the
\href{https://github.com/Hironsan/ja.text8}{ja.text8} corpus as a tibble
by splitting it into sentences.
The ja.text8 as whole corpus consists of over 582,000 sentences,
16,900,026 tokens, and 290,811 vocabularies.
}
