% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrte-reader.R
\name{read_jrte_rte}
\alias{read_jrte_rte}
\title{JRTE Coprus Parser}
\usage{
read_jrte_rte(
  url = "https://github.com/megagonlabs/jrte-corpus/archive/refs/heads/master.zip",
  exdir = tempdir(),
  keep = jrte_rte_files()
)
}
\arguments{
\item{url}{String.}

\item{exdir}{String. Path to tempolarily unzip text files.}

\item{keep}{List. File names to parse and keep in data.frame.}
}
\value{
list of data.frames
}
\description{
Download and read the Japanese Realistic Textual Entailment Corpus.
The result of this function is memoised with \code{memoise::memoise} internally.
}
