% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldnws-reader.R
\name{read_ldnws}
\alias{read_ldnws}
\title{Read the Livedoor News Corpus}
\usage{
read_ldnws(
  url = "https://www.rondhuit.com/download/ldcc-20140209.tar.gz",
  exdir = tempdir(),
  keep = ldnws_categories(),
  collapse = "\\n\\n",
  include_title = TRUE
)
}
\arguments{
\item{url}{String.
If left with \code{NULL}, the function will skip downloading the file.}

\item{exdir}{String. Directory to tempolarily untar text files.}

\item{keep}{Categories to parse and keep in the tibble.}

\item{collapse}{String with which \code{\link[base:paste]{base::paste()}} collapses lines.}

\item{include_title}{Logical. Whether to include title in text body field.
Defaults to \code{TRUE}.}
}
\value{
A tibble.
}
\description{
Downloads and reads the Livedoor News Corpus.
The result of this function is memoised
with \code{\link[memoise:memoise]{memoise::memoise()}} internally.
}
\details{
This function downloads the Livedoor News Corpus and parses it to a tibble.
For details about the Livedoor News Corpus, please see
\href{https://www.rondhuit.com/download.html#ldcc}{thie page}.
}
\seealso{
Other ldnws-reader:
\code{\link[=ldnws_categories]{ldnws_categories()}}
}
\concept{ldnws-reader}
