% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldccr-package.R
\name{read_srt}
\alias{read_srt}
\title{Read srt file}
\usage{
read_srt(path, collapse = "\\n")
}
\arguments{
\item{path}{Path to srt file.}

\item{collapse}{String.}
}
\value{
A tibble.
}
\description{
Read srt file
}
