% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sqids_encode}
\alias{sqids_encode}
\alias{sqids_decode}
\title{Encode and decode 'Sqids'}
\usage{
sqids_encode(numbers, salt)

sqids_decode(ids)
}
\arguments{
\item{numbers}{A list of integers.}

\item{salt}{An integer vector.}

\item{ids}{A character vector.}
}
\value{
For \code{sqids_encode()}, a list of integers.
For \code{sqids_decode()}, a character vector.
}
\description{
Encode and decode 'Sqids'
}
\keyword{internal}
