% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unidic-downloader.R
\name{unidic_availables}
\alias{unidic_availables}
\alias{download_unidic}
\title{Download and unzip 'UniDic'}
\usage{
unidic_availables()

download_unidic(version = "latest", dirname = "unidic")
}
\arguments{
\item{version}{String; version of 'UniDic'.}

\item{dirname}{String; directory to unzip the dictionary.}
}
\value{
Full path to \code{dirname} is returned invisibly.
}
\description{
Downloads 'UniDic' of specified version into \code{dirname}.
This function is a partial port of
\href{https://github.com/polm/unidic-py}{polm/unidic-py}.
Note that to unzip dictionary will take up 770MB on disk after downloading.
}
