% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{clean_url}
\alias{clean_emoji}
\title{Utility functions}
\usage{
clean_url(text, replacement = "")

clean_emoji(text, replacement = "")
}
\arguments{
\item{text}{A character vector.}

\item{replacement}{String.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
These functions are experimental and may withdraw in the future.
}
