% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldccr-package.R
\docType{package}
\name{ldccr-package}
\alias{ldccr}
\alias{ldccr-package}
\title{ldccr: Utilities for Various Japanese Corpora}
\description{
The goal of ldccr package is to make easy to use Japanese language resources. This package provides parsers for several Japanese corpora that are free or open licensed and a downloader of zipped text files published on Aozora Bunko.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/ldccr}
  \item Report bugs at \url{https://github.com/paithiov909/ldccr/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com} (2023-present Sqids maintainers [cph])

}
\keyword{internal}
