% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldccr-package.R
\docType{package}
\name{ldccr-package}
\alias{ldccr}
\alias{ldccr-package}
\title{Various Japanese Corpus Parser}
\description{
Various Japanese corpus parser for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/ldccr}
  \item Report bugs at \url{https://github.com/paithiov909/ldccr/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
