% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldnws-reader.R
\name{ldnws_categories}
\alias{ldnws_categories}
\title{List of categories of the Livedoor News Corpus}
\usage{
ldnws_categories(
  keep = c("dokujo-tsushin", "it-life-hack", "kaden-channel", "livedoor-homme",
    "movie-enter", "peachy", "smax", "sports-watch", "topic-news")
)
}
\arguments{
\item{keep}{Character vector. Category names to parse.}
}
\value{
A character vector.
}
\description{
List of categories of the Livedoor News Corpus
}
