% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldnws-reader.R
\docType{data}
\name{ldnws_categories}
\alias{ldnws_categories}
\title{List of Categories}
\format{
An object of class \code{list} of length 9.
}
\usage{
ldnws_categories
}
\value{
list
}
\description{
List of Categories
}
\keyword{internal}
