% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrte-parse.R
\name{parse_jrte_judges}
\alias{parse_jrte_judges}
\title{Parse judges column of 'rte.*.tsv'}
\usage{
parse_jrte_judges(tbl)
}
\arguments{
\item{tbl}{A tibble returned from \code{read_jrte} which name is \code{rte.*.tsv}
(this function cannot parse judges of the \code{pn} dataset).}
}
\value{
tibble
}
\description{
Parse judges column of 'rte.*.tsv'
}
\keyword{internal}
