% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrte-parse.R
\name{parse_jrte_reasoning}
\alias{parse_jrte_reasoning}
\title{Parse reasoning column of 'rte.*.tsv'}
\usage{
parse_jrte_reasoning(tbl)
}
\arguments{
\item{tbl}{A tibble returned from \code{\link[=read_jrte]{read_jrte()}} of which name is \code{rte.*.tsv}.}
}
\value{
A tibble.
}
\description{
Parse reasoning column of 'rte.*.tsv'
}
\seealso{
Other jrte-reader: 
\code{\link{jrte_rte_files}()},
\code{\link{parse_jrte_judges}()},
\code{\link{read_jrte}()}
}
\concept{jrte-reader}
