% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrte-parse.R
\name{parse_jrte_reasoning}
\alias{parse_jrte_reasoning}
\title{Parse reasoning column of 'rte.*.tsv'}
\usage{
parse_jrte_reasoning(tbl)
}
\arguments{
\item{tbl}{A tibble returned from \code{read_jrte} which name is \code{rte.*.tsv}.}
}
\value{
tibble
}
\description{
Parse reasoning column of 'rte.*.tsv'
}
