% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aozora-reader.R
\name{read_aozora}
\alias{read_aozora}
\title{Download text file from Aozora Bunko}
\usage{
read_aozora(
  url = "https://www.aozora.gr.jp/cards/000081/files/472_ruby_654.zip",
  txtname = NULL,
  directory = file.path(getwd(), "cache")
)
}
\arguments{
\item{url}{URL of text download link.}

\item{txtname}{New file name as which text is saved.
If left to \code{NULL}, keeps name of the source file.}

\item{directory}{Path where new file is saved.}
}
\value{
The path to the file downloaded.
}
\description{
Download a file from specified URL, unzip and convert it to UTF-8.
}
