% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrte-reader.R
\name{read_jrte}
\alias{read_jrte}
\title{JRTE Coprus Parser}
\usage{
read_jrte(
  url = "https://github.com/megagonlabs/jrte-corpus/archive/refs/heads/master.zip",
  exdir = tempdir(),
  keep = jrte_rte_files(),
  keep_rhr = FALSE,
  keep_pn = FALSE
)
}
\arguments{
\item{url}{String.}

\item{exdir}{String. Path to tempolarily unzip text files.}

\item{keep}{List. File names to parse and keep in returned value.}

\item{keep_rhr}{Logical. If true, function keeps \code{rhr.tsv}.}

\item{keep_pn}{Logical. If true, function keeps \code{pn.tsv}.}
}
\value{
list of tibbles
}
\description{
Download and read the Japanese Realistic Textual Entailment Corpus.
The result of this function is memoised with \code{memoise::memoise} internally.
}
