% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldnws-reader.R
\name{read_ldnws}
\alias{read_ldnws}
\title{Read the Livedoor News Corpus}
\usage{
read_ldnws(
  url = "https://www.rondhuit.com/download/ldcc-20140209.tar.gz",
  exdir = tempdir(),
  keep = ldnws_categories(),
  collapse = "\\n\\n"
)
}
\arguments{
\item{url}{String.}

\item{exdir}{String. Path to tempolarily untar text files.}

\item{keep}{List. Categories to parse and keep in data.frame.}

\item{collapse}{String with which \code{base::paste} collapses lines.}
}
\value{
tibble
}
\description{
Download and read the Livedoor News Corpus.
The result of this function is memoised with \code{memoise::memoise} internally.
}
\details{
This function downloads the Livedoor News Corpus and parses it to a tibble.
For details about the Livedoor News Corpus, please see
\href{https://www.rondhuit.com/download.html#ldcc}{this page}.
}
